+++ qio/src/org/qcmg/unused/BEDFileReader.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.genesymbol;
+package org.qcmg.unused;
 
 import java.io.Closeable;
 import java.io.File;
@@ -10,26 +10,25 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.util.Iterator;
 
-public final class GeneSymbolFileReader implements Closeable, Iterable<GeneSymbolRecord> {
+public final class BEDFileReader implements Closeable, Iterable<BEDRecord> {
     private final File file;
     private final InputStream inputStream;
 
-    public GeneSymbolFileReader(final File file) throws IOException {
+    public BEDFileReader(final File file) throws IOException {
         this.file = file;
         FileInputStream fileStream = new FileInputStream(file);
         inputStream = fileStream;
     }
 
-    public Iterator<GeneSymbolRecord> iterator() {
+    public Iterator<BEDRecord> iterator() {
         return getRecordIterator();
     }
 
-    public GeneSymbolRecordIterator getRecordIterator() {
-        return new GeneSymbolRecordIterator(inputStream);
+    public BEDRecordIterator getRecordIterator() {
+        return new BEDRecordIterator(inputStream);
     }
 
     public void close() throws IOException {
-    	inputStream.close();
     }
 
     public File getFile() {
