+++ qio/src/org/qcmg/unused/BEDRecordIterator.java
@@ -1,10 +1,7 @@
 /**
- * © Copyright The University of Queensland 2010-2014.
- * © Copyright QIMR Berghofer Medical Research Institute 2014-2016.
- *
- * This code is released under the terms outlined in the included LICENSE file.
+ * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.maf;
+package org.qcmg.unused;
 
 import java.io.BufferedReader;
 import java.io.InputStream;
@@ -12,13 +9,11 @@ import java.io.InputStreamReader;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
 
-import org.qcmg.common.maf.MAFRecord;
-
-public final class MAFRecordIterator implements Iterator<MAFRecord> {
+public final class BEDRecordIterator implements Iterator<BEDRecord> {
     private final BufferedReader reader;
-    private MAFRecord next;
+    private BEDRecord next;
 
-    public MAFRecordIterator(final InputStream stream) {
+    public BEDRecordIterator(final InputStream stream) {
         InputStreamReader streamReader = new InputStreamReader(stream);
         reader = new BufferedReader(streamReader);
         readNext();
@@ -28,18 +23,18 @@ public final class MAFRecordIterator implements Iterator<MAFRecord> {
         return null != next;
     }
 
-    public MAFRecord next() {
+    public BEDRecord next() {
         if (!hasNext()) {
             throw new NoSuchElementException();
         }
-        MAFRecord result = next;
+        BEDRecord result = next;
         readNext();
         return result;
     }
 
     private void readNext() {
         try {
-            next = MAFSerializer.nextRecord(reader);
+            next = BEDSerializer.nextRecord(reader);
         } catch (NoSuchElementException e) {
             throw e;
         } catch (Exception ex) {
