+++ qio/src/org/qcmg/unused/GeneSymbolRecord.java
@@ -1,7 +1,7 @@
 /**
  * © Copyright The University of Queensland 2010-2014.  This code is released under the terms outlined in the included LICENSE file.
  */
-package org.qcmg.genesymbol;
+package org.qcmg.unused;
 
 
 public class GeneSymbolRecord {
