+++ qmule/src/org/qcmg/qmule/io/ChrConvRecordIterator.java
