+++ qmule/src/org/qcmg/qmule/io/ChrConvSerializer.java
