+++ qmule/src/org/qcmg/qmule/io/VerifiedSnpFileReader.java
