+++ qmule/src/org/qcmg/qmule/io/VerifiedSnpRecord.java
