+++ qmule/src/org/qcmg/qmule/io/VerifiedSnpRecordIterator.java
