+++ qmule/src/org/qcmg/qmule/io/VerifiedSnpSerializer.java
